+++ src/main/java/domain/card/CardBundle.java
@@ -1,8 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.card.Card;
-import domain.card.Rank;
-import domain.card.Shape;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
