+++ src/main/java/domain/card/CardDeck.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.card.Card;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.List;
