+++ src/main/java/domain/participant/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import java.util.List;
