+++ src/test/java/domain/BlackJackGameTest.java
@@ -4,9 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import domain.card.Card;
+import domain.card.CardBundle;
+import domain.card.CardDeck;
 import domain.card.Rank;
 import domain.card.Shape;
 import domain.fixture.ParticipantsFixture;
+import domain.participant.Participant;
+import domain.participant.Participants;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
