+++ src/test/java/domain/card/CardDeckTest.java
@@ -1,10 +1,7 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.card.Card;
-import domain.card.Rank;
-import domain.card.Shape;
 import java.util.Deque;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
